Modification to chem/Makefile (not to chem/Makefile_org)
When modifications are necessary to chem/Makefile,
it should be done to chem/Makefile_org (and not chem/Makefile).
But in the WRF git archive, chem/Makefile_org is not included.
During the compilation, chem/Makefile_org is copied from chem/Makefile
if chem/Makefile_org does not exist.
chem/Makefile_org is also deleted when "./clean -aa" is run.
To avoid losing modifications made to chem/Makefile_org,
In this commit, all modifications are done to chem/Makefile.
Therefore, it is necessary to rerun './clean -aa' or simply delete
chem/Makefile_org, if it exists.